+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/stories/Story.kt
@@ -1,0 +1,6 @@
+package dev.inmo.tgbotapi.types.stories
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+class Story