+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # TelegramBotAPI changelog
 
+## 9.1.0
+
+**This update contains adding of [Telegram Bot API 6.8](https://core.telegram.org/bots/api-changelog#august-18-2023) support**
+
+* `Version`:
+  * `Coroutines`: `1.7.2` -> `1.7.3`
+  * `Ktor`: `2.3.2` -> `2.3.3`
+  * `MicroUtils`: `0.19.7` -> `0.19.9`
+
 ## 9.0.0
 
 **THIS UPDATE CONTAINS BREAKING CHANGES: USERNAMES OF BOTS NOW BECAME NULLABLE**